+++ src/main/java/com/newdeal/ledger/card/dto/CardType.java
@@ -1,0 +1,5 @@
+package com.newdeal.ledger.card.dto;
+
+public enum CardType {
+	CREDIT, CHECK
+}
