+++ src/main/java/com/newdeal/ledger/card/mapper/CardMapper.java
@@ -1,20 +1,21 @@
-package com.newdeal.ledger.cardaccount.mapper;
+package com.newdeal.ledger.card.mapper;
 
 import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
-import com.newdeal.ledger.cardaccount.dto.CardDto;
-import com.newdeal.ledger.cardaccount.dto.CardRequest;
+import com.newdeal.ledger.card.dto.CardDto;
+import com.newdeal.ledger.card.dto.CardRequest;
 
 @Mapper
 public interface CardMapper {
-	List<CardDto> findAllByEmail(String email);
-
 	void createCard(@Param("email") String email, @Param("request") CardRequest.Create request);
 
 	void updateCard(@Param("cardId") Integer cardId, @Param("request") CardRequest.Update request);
 
 	void deleteCard(Integer cardId);
+
+	List<CardDto> findCardsByEmail(String email);
+
 }
