--- src/main/java/com/newdeal/ledger/cardaccount/dto/type/CardType.java
@@ -1,5 +1,0 @@
-package com.newdeal.ledger.cardaccount.dto.type;
-
-public enum CardType {
-	CREDIT, CHECK
-}
