--- src/main/java/com/newdeal/ledger/cardaccount/service/CardService.java
@@ -1,16 +1,0 @@
-package com.newdeal.ledger.cardaccount.service;
-
-import java.util.List;
-
-import com.newdeal.ledger.cardaccount.dto.CardDto;
-import com.newdeal.ledger.cardaccount.dto.CardRequest;
-
-public interface CardService {
-	List<CardDto> selectAllByEmail(String email);
-
-	void createCard(String tempEmail, CardRequest.Create request);
-
-	void updateCard(Integer cardId, CardRequest.Update request);
-
-	void deleteCard(Integer cardId);
-}
